+++ .idea/misc.xml
@@ -4,6 +4,7 @@
     <option name="filePathToZoomLevelMap">
       <map>
         <entry key="..\:/Users/Xiaomi/Desktop/Android/RoomHomeWork/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.198" />
+        <entry key="..\:/Users/Xiaomi/Desktop/Android/RoomHomeWork/app/src/main/res/drawable/ic_history_background.xml" value="0.221" />
         <entry key="..\:/Users/Xiaomi/Desktop/Android/RoomHomeWork/app/src/main/res/drawable/ic_launcher_background.xml" value="0.198" />
         <entry key="..\:/Users/Xiaomi/Desktop/Android/RoomHomeWork/app/src/main/res/layout/activity_main.xml" value="0.24010416666666667" />
         <entry key="..\:/Users/Xiaomi/Desktop/Android/RoomHomeWork/app/src/main/res/layout/fragment_history.xml" value="0.25416666666666665" />
