+++ app/src/main/java/com/example/roomhomework/presentation/di/viewModelModule.kt
@@ -1,9 +1,11 @@
 package com.example.roomhomework.presentation.di
 
-import com.example.roomhomework.presentation.viewModel.OrdersViewModel
+import com.example.roomhomework.presentation.viewModels.HistoryViewModel
+import com.example.roomhomework.presentation.viewModels.OrdersViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
 val viewModelModule = module {
     viewModel{OrdersViewModel(interactor = get())}
+    viewModel{HistoryViewModel(interactor = get())}
 }
