+++ app/src/main/java/com/example/roomhomework/presentation/fragments/RegistrationFragment.kt
@@ -4,8 +4,9 @@ package com.example.roomhomework.presentation.fragments
 import androidx.fragment.app.Fragment
 import com.example.roomhomework.R
 import com.example.roomhomework.data.storage.entity.PeopleAge
-import com.example.roomhomework.presentation.viewModel.OrdersViewModel
+import com.example.roomhomework.presentation.viewModels.OrdersViewModel
 import com.example.roomhomework.utils.openFragment
+import com.example.roomhomework.utils.showSnack
 import kotlinx.android.synthetic.main.fragment_registration.*
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
@@ -26,27 +27,44 @@ class RegistrationFragment : Fragment(R.layout.fragment_registration) {
 
     private fun createOrder() {
         btnConfirm.setOnClickListener {
-            val type =
-                viewModel.viewSwitchType(
-                    viewSwitch,
-                    getString(PeopleAge.CHILD.type),
-                    getString(PeopleAge.ADULT.type))
-            viewModel.createOrder(
-                fromTown = fromTown.text.toString(),
-                toTown = toTown.text.toString(),
-                departmentTime = departureTime.text.toString(),
-                arriveTime = arrivalTime.text.toString(),
-                personName = name.text.toString(),
-                password = password.text.toString(),
-                age = type,
-                view = requireView(),
-                messageEmpty = getString(R.string.snack_empty),
-                messageDate = getString(R.string.snack_date),
-                messagePassword = getString(R.string.snack_password)
-
-            )
+            val fromTown = fromTown.text.toString()
+            val toTown = toTown.text.toString()
+            val departmentTime = departureTime.text.toString()
+            val arriveTime = arrivalTime.text.toString()
+            val personName = name.text.toString()
+            val password = password.text.toString()
+            when (viewModel.validation(
+                fromTown,
+                toTown,
+                departmentTime,
+                arriveTime,
+                personName,
+                password
+            )) {
+                OrdersViewModel.EMPTY_VALIDATION -> showSnack(getString(R.string.snack_empty))
+                OrdersViewModel.DATE_VALIDATION -> showSnack(getString(R.string.snack_date))
+                OrdersViewModel.PASSWORD_VALIDATION -> showSnack(getString(R.string.snack_password))
+                else -> {
+                    val type =
+                        viewModel.viewSwitchType(
+                            viewSwitch,
+                            getString(PeopleAge.CHILD.type),
+                            getString(PeopleAge.ADULT.type)
+                        )
+                    viewModel.createOrder(
+                        fromTown,
+                        toTown,
+                        departmentTime,
+                        arriveTime,
+                        personName,
+                        password,
+                        type
+                    )
+                }
+            }
         }
     }
+
     private fun openHistoryFragment() {
         btnHistory.setOnClickListener {
             requireActivity().openFragment(
