+++ app/src/main/res/layout/fragment_registration.xml
@@ -140,7 +140,8 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:hint="@string/password"
-            android:textColorHint="@color/hint_color" />
+            android:textColorHint="@color/hint_color"
+            android:maxLength="@integer/password_count"/>
     </com.google.android.material.textfield.TextInputLayout>
 
     <androidx.constraintlayout.widget.ConstraintLayout
